requirements.txt
streamlit
scikit-learn
nltk
sentence-transformers
PyPDF2
python-docx
